+++ src/main.c
@@ -3,7 +3,7 @@
 
 #include "main.h"
 #include "thread0_epc.h"
-#include "thread3_main.h"
+#include "unknown_06B2B0.h"
 #include "lib/src/libc/rmonPrintf.h"
 
 /************ .rodata ************/
