+++ src/unknown_00BC20.c
@@ -12,7 +12,7 @@
 #include "thread0_epc.h"
 #include "controller_pak.h"
 #include "menu.h"
-#include "thread3_main.h"
+#include "unknown_06B2B0.h"
 #include "unknown_043920.h"
 
 /************ .data ************/
