+++ src/unknown_043920.c
@@ -15,7 +15,7 @@
 #include "unknown_00BC20.h"
 #include "audio.h"
 #include "unknown_0348C0.h"
-#include "thread3_main.h"
+#include "unknown_06B2B0.h"
 
 #define MAX_NUMBER_OF_GHOST_NODES 360
